+++ src/main/java/assessment/parkinglot/controller/VehicleToPark.java
@@ -1,0 +1,9 @@
+package assessment.parkinglot.controller;
+
+public class VehicleToPark {
+    private String type;
+
+    public String getType() {
+        return type;
+    }
+}