+++ src/main/java/assessment/parkinglot/enums/ParkingLotAllowedTypes.java
@@ -1,0 +1,8 @@
+package assessment.parkinglot.enums;
+
+public enum ParkingLotAllowedTypes {
+    MOTORCYCLE,
+    CAR,
+    VAN,
+    ;
+}