+++ src/main/java/assessment/parkinglot/enums/ParkingLotSpots.java
@@ -1,0 +1,8 @@
+package assessment.parkinglot.enums;
+
+public enum ParkingLotSpots {
+    MOTORCYCLE_SPOT,
+    COMPACT_CAR_SPOT,
+    REGULAR_SPOT,
+    ;
+}