+++ src/main/java/assessment/parkinglot/exceptions/ParkingVehicleException.java
@@ -1,0 +1,7 @@
+package assessment.parkinglot.exceptions;
+
+public class ParkingVehicleException extends Exception {
+    public ParkingVehicleException(String message) {
+        super(message);
+    }
+}