+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 # Changelog
 
-## [0.0.1] - 2024-02-22 14:47
+### TODO
+- Repository
+- Vehicle class
+- Exceptions
+- Handle exceptions (ControllerAdvice)
+- Classes info
+- Serialize/Deserialize objects
+- Logger
+- Unit testing
+  - Controller
+  - Service
+  - Repository
 
+## [0.0.1] - 2024-02-22 14:47
+### Added
+- Controller, Service
+- Validation
+- POST mapping
+- Exceptions
+- Steps to perform the task
+
+## [0.0.1] - 2024-02-22 14:47
 ### Added
 - First commit
 - Zip file added to github repository